Pull quicksort into separate file
I will eventually add unit tests & fix any bugs
